IDPROD-2953 part 4: Document Uploader (#630)
New `DocumentUploader` coordinator class to
manage uploading all of the required identity
document images at the correct sizes and jpeg
qualities. The image upload logic was refactored
out of `DocumentCaptureViewController` and into
`DocumentUploader` so it could easily be reused
for the case the user needs to upload an image
from their camera roll (not implemented yet).

+++ StripeIdentity/StripeIdentity/Source/API Bindings/Models/VerificationPage/VerificationPageStaticContentDocumentCapturePage.swift
@@ -17,7 +17,7 @@ struct VerificationPageStaticContentDocumentCapturePage: StripeDecodable, Equata
     let highResImageCropPadding: Decimal
     let highResImageMaxDimension: Int
     let lowResImageCompressionQuality: Decimal
-    let lowResImageMaxDimension: Decimal
+    let lowResImageMaxDimension: Int
     let models: VerificationPageStaticContentDocumentCaptureModels
     let requireLiveCapture: Bool
